move adapter an descriptors into separate classes for an easier migration, also add first step to convert to compose adapter
